Refactoring: move file storage to separate package
