+++ logic/src/main/java/org/itsallcode/whiterabbit/logic/storage/CachingStorage.java
@@ -1,23 +1,24 @@
 package org.itsallcode.whiterabbit.logic.storage;
 
-import org.itsallcode.whiterabbit.logic.model.DayRecord;
-import org.itsallcode.whiterabbit.logic.service.contract.ContractTermsService;
-import org.itsallcode.whiterabbit.logic.service.project.ProjectService;
+import java.nio.file.Path;
+import java.time.LocalDate;
+import java.util.List;
 
 import javax.json.bind.Jsonb;
 import javax.json.bind.JsonbBuilder;
 import javax.json.bind.JsonbConfig;
-import java.nio.file.Path;
-import java.time.LocalDate;
-import java.util.List;
+
+import org.itsallcode.whiterabbit.logic.model.DayRecord;
+import org.itsallcode.whiterabbit.logic.service.contract.ContractTermsService;
+import org.itsallcode.whiterabbit.logic.service.project.ProjectService;
+import org.itsallcode.whiterabbit.logic.storage.data.JsonFileStorage;
 
 public interface CachingStorage extends Storage
 {
     static CachingStorage create(Path dataDir, ContractTermsService contractTerms, ProjectService projectService)
     {
-        final DateToFileMapper dateToFileMapper = new DateToFileMapper(dataDir);
         final Jsonb jsonb = JsonbBuilder.create(new JsonbConfig().withFormatting(true));
-        final JsonFileStorage fileStorage = new JsonFileStorage(jsonb, dateToFileMapper);
+        final JsonFileStorage fileStorage = new JsonFileStorage(jsonb, dataDir);
         final MonthIndexStorage monthIndexStorage = new MonthIndexStorage(contractTerms, projectService, fileStorage);
         return new CachingStorageImpl(monthIndexStorage);
     }
