+++ logic/src/test/java/org/itsallcode/whiterabbit/logic/storage/MonthIndexStorageTest.java
@@ -51,7 +51,7 @@ class MonthIndexStorageTest
     @Test
     void loadPreviousMonthOvertime_returnsZero_whenNoPreviousMonth()
     {
-        when(fileStorageMock.loadMonthRecord(YearMonth.of(2020, Month.OCTOBER))).thenReturn(Optional.empty());
+        when(fileStorageMock.loadMonth(YearMonth.of(2020, Month.OCTOBER))).thenReturn(Optional.empty());
 
         assertThat(storage.loadPreviousMonthOvertime(YEAR_MONTH)).isZero();
     }
@@ -59,7 +59,7 @@ class MonthIndexStorageTest
     @Test
     void loadPreviousMonthOvertime_returnsPreviousMonthOvertime()
     {
-        when(fileStorageMock.loadMonthRecord(PREVIOUS_MONTH))
+        when(fileStorageMock.loadMonth(PREVIOUS_MONTH))
                 .thenReturn(Optional.of(jsonMonth(PREVIOUS_MONTH, Duration.ofMinutes(5))));
 
         assertThat(storage.loadPreviousMonthOvertime(YEAR_MONTH)).hasMinutes(5);
@@ -68,7 +68,7 @@ class MonthIndexStorageTest
     @Test
     void loadPreviousMonthOvertime_truncatesToMinute()
     {
-        when(fileStorageMock.loadMonthRecord(PREVIOUS_MONTH))
+        when(fileStorageMock.loadMonth(PREVIOUS_MONTH))
                 .thenReturn(Optional.of(jsonMonth(PREVIOUS_MONTH, Duration.ofMinutes(5).plusSeconds(50))));
 
         assertThat(storage.loadPreviousMonthOvertime(YEAR_MONTH)).hasMinutes(5);
@@ -77,7 +77,7 @@ class MonthIndexStorageTest
     @Test
     void loadOrCreate_createsNewMonthIfNotExists()
     {
-        when(fileStorageMock.loadMonthRecord(YEAR_MONTH)).thenReturn(Optional.empty());
+        when(fileStorageMock.loadMonth(YEAR_MONTH)).thenReturn(Optional.empty());
 
         final MonthIndex newMonth = storage.loadOrCreate(YEAR_MONTH);
 
@@ -91,7 +91,7 @@ class MonthIndexStorageTest
     void loadOrCreate_loadsExistingMonth()
     {
         final JsonMonth month = jsonMonth(YEAR_MONTH, Duration.ofMinutes(4));
-        when(fileStorageMock.loadMonthRecord(YEAR_MONTH)).thenReturn(Optional.of(month));
+        when(fileStorageMock.loadMonth(YEAR_MONTH)).thenReturn(Optional.of(month));
 
         final MonthIndex newMonth = storage.loadOrCreate(YEAR_MONTH);
 
@@ -111,7 +111,7 @@ class MonthIndexStorageTest
 
         storage.storeMonth(monthIndexMock);
 
-        verify(fileStorageMock).writeToFile(eq(YEAR_MONTH), same(month));
+        verify(fileStorageMock).store(eq(YEAR_MONTH), same(month));
     }
 
     @Test
